requirements.txt
aiomysql==0.0.21
aioredis==1.3.1
async-timeout==3.0.1
click==7.1.2
databases==0.4.1
fastapi==0.63.0
h11==0.12.0
hiredis==1.1.0
numpy==1.19.5
pandas==1.2.0
passlib==1.7.4
pydantic==1.7.3
PyJWT==2.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.3
yapf==0.30.0
